+++ injectable_generator/lib/dependency_config.dart
@@ -48,12 +48,6 @@ class DependencyConfig {
         ...imports.where((i) => i != null),
         ...dependencies.map((dep) => dep.import).where((i) => i != null),
       };
-
-  bool dependsOn(DependencyConfig b) {
-    final depends = dependencies.map((d) => d.type).toSet().contains(b.type);
-    // print('$type depends($depends) on ${b.type}');
-    return depends;
-  }
 }
 
 class InjectedDependency {