+++ injectable_generator/lib/injectable_generator.dart
@@ -54,9 +54,12 @@ class InjectableGenerator implements Generator {
   }
 
   bool _hasConventionalMatch(ClassElement clazz) {
+    if (clazz.isAbstract) {
+      return false;
+    }
     final fileName = clazz.source.shortName.replaceFirst('.dart', '');
     return (_classNameMatcher != null &&
-            _classNameMatcher.hasMatch(clazz.name)) ||
+        _classNameMatcher.hasMatch(clazz.name)) ||
         (_fileNameMatcher != null && _fileNameMatcher.hasMatch(fileName));
   }
 }